backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic-settings
python-dotenv
email-validator
httpx
joblib
numpy
scikit-learn
